Fixed caching implementation; Reorganized the application and its methods into their respective microservices;
