+++ routes/console.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\StockPriceController;
+use App\Jobs\FetchApiData;
 use Illuminate\Foundation\Inspiring;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Schedule;
@@ -9,8 +9,5 @@ Artisan::command('inspire', function () {
     $this->comment(Inspiring::quote());
 })->purpose('Display an inspiring quote')->hourly();
 
-Schedule::call(function () {
-    $stockPriceController = app(StockPriceController::class);
-    $stockPrices = $stockPriceController->fetchStockPrices();
-    dump($stockPrices);
-})->everyMinute();
+$fetchApiData = app(FetchApiData::class);
+Schedule::job($fetchApiData)->everyMinute();
